Added ToImage method and supporting JS functions
The `Chart` class in `Chart.razor.cs` has been extended with a new asynchronous method `ToImage()`, which invokes a JavaScript function `toImage` with the `Id` of the chart. This function is expected to return an image of the chart as a byte array. In `Chart.razor.js`, two new functions have been added: `canvasToBlob(canvas)` and `toImage(id)`. The former takes a canvas element and returns a Promise that resolves with a byte array representing the image data of the canvas. The latter takes an id, finds the HTML element with the given id, retrieves the first canvas within this element, and uses the `canvasToBlob` function to get the image data.