`Update image conversion methods to accept mimeType parameter`
Updated the `ToImage` method in `Chart.razor.cs` to `ToImageAsync` and added an optional `mimeType` parameter with a default value of `"image/png"`. This parameter is passed to the `InvokeAsync` method.

In `Chart.razor.js`, the `canvasToBlob` and `toImage` functions now accept a `mimeType` parameter, which is passed to the `canvas.toBlob` method and `canvasToBlob` function respectively. The error handling for the `FileReader` has been improved to provide a more descriptive error message. The `toImage` function continues to return a promise that resolves with a blob array or rejects with an error.