Add `customCanvasBackgroundColor` plugin to `Chart.razor.js`
Added a new plugin named `customCanvasBackgroundColor` to `Chart.razor.js` that allows setting a custom background color for the chart canvas. The `chartOption` object and `getChartOption` function were updated to include this new plugin. Also, a new property `CanvasBackgroundColor` was added to the `ChartOptions` class in `ChartOptions.cs` to get or set the background color of the chart canvas.